2.0. ffmpeg_producer: Fixed integer overflow when looping.      image_mixer: Refactored draw call parameters.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@1282 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ modules/ffmpeg/producer/ffmpeg_producer.cpp
@@ -172,7 +172,7 @@ public:
 	virtual int64_t nb_frames() const 
 	{
 		if(loop_)
-			return std::numeric_limits<int64_t>::max();
+			return std::numeric_limits<int>::max();
 
 		// This function estimates nb_frames until input has read all packets for one loop, at which point the count should be accurate.
